+++ README.md
@@ -1,1 +1,8 @@
-readme!
+# 📝Paper📝
+
+## 📆개발기간
+
+2022.06.24 - 2022.08.06
+
+## 내용
+
