+++ src/components/main/Header.jsx
@@ -8,8 +8,11 @@ import defaultUserImage from "../../public/images/default_profile.png";
 import Swal from "sweetalert2";
 /* 컴포넌트 */
 import HeadPaperSearch from "./HeadPaperSearch";
+import { BiSearchAlt2 } from "react-icons/bi";
+
 
 const Header = () => {
+  const navigate = useNavigate();
   const onLogout = () => {
     deleteCookie("token");
     deleteCookie("nickname");
@@ -27,7 +30,7 @@ const Header = () => {
     });
     navigate("/");
   };
-  const navigate = useNavigate();
+
   /* 쿠키 */
   const cookie = getCookie("token");
   const blogId = getCookie("blogId");
@@ -65,23 +68,31 @@ const Header = () => {
     <>
       <HeaderBox>
         <Svg>
-          <Logo>
-            <Link to="/" style={{ color: "inherit", textDecoration: "none" }}>
-              <img
-                className="paperTitle"
-                src={process.env.PUBLIC_URL + "/Frame.png"}
-                back_size="100% 100%"
-                alt="icon"
-              />
-            </Link>
-          </Logo>
-          <Search>
-            <HeadPaperSearch />
-          </Search>
+
+
+          <img
+            onClick={() => {
+              navigate("/");
+            }}
+            className="paperTitle"
+            src={process.env.PUBLIC_URL + "/logo_paper.png"}
+            back_size="100% 100%"
+            alt="icon"
+          />
+
+
           <Login>
             {is_cookie ? (
               <>
-                <BtnBox>
+                <BtnBox1>
+                  <SearchIcon1
+                    onClick={() => {
+                      navigate(`/paper/search`);
+                    }}
+                  >
+                    <BiSearchAlt2 color="black" size="25px" />
+                  </SearchIcon1>
+
                   <BtnItem>
                     <Btn
                       onClick={() => {
@@ -91,65 +102,72 @@ const Header = () => {
                       전체글
                     </Btn>
                   </BtnItem>
-                  <BtnItem>
-                    <Btn
-                      onClick={() => {
-                        navigate("/write");
-                      }}
-                    >
-                      작성하기
-                    </Btn>
-                  </BtnItem>
-                  <DropDownContainer ref={el}>
-                    <ProfileImgBox
-                      src={
-                        profileImage === "null" || profileImage === "undefined"
-                          ? defaultUserImage
-                          : S3
-                      }
-                      onClick={toggling}
-                    />
-                    {isOpen && (
-                      <DropDownListContainer>
-                        <DropDownList>
-                          <ListItem
-                            onClick={() => {
-                              navigate(`/paper/${blogId}`);
-                            }}
-                          >
-                            내 블로그
-                          </ListItem>
-                          <ListItem
-                            onClick={() => {
-                              navigate(`/myprofile`);
-                            }}
-                          >
-                            회원정보
-                          </ListItem>
-                          <ListItem
-                            onClick={() => {
-                              navigate(`/paper/${blogId}/reservationList`);
-                            }}
-                          >
-                            예약리스트
-                          </ListItem>
-                          <ListItem
-                            onClick={() => {
-                              onLogout();
-                            }}
-                          >
-                            로그아웃
-                          </ListItem>
-                        </DropDownList>
-                      </DropDownListContainer>
-                    )}
-                  </DropDownContainer>
-                </BtnBox>
-                <NickBox>{nickname}</NickBox>
+                </BtnBox1>
+                <DropDownContainer ref={el}>
+                  <ProfileImgBox
+                    src={
+                      profileImage === "null" || profileImage === "undefined"
+                        ? defaultUserImage
+                        : S3
+                    }
+                    onClick={toggling}
+                  />
+                  {isOpen && (
+                    <DropDownListContainer>
+                      <DropDownList>
+                        <ListItem
+                          onClick={() => {
+                            navigate(`/paper/${blogId}`);
+                          }}
+                        >
+                          내 블로그
+                        </ListItem>
+                        <ListItem
+                          onClick={() => {
+                            navigate(`/paper/${blogId}/reservationList`);
+                          }}
+                        >
+                          예약리스트
+                        </ListItem>
+                        <ListItem
+                          onClick={() => {
+                            navigate(`/write`);
+                          }}
+                        >
+                          글 작성하기
+                        </ListItem>
+                        <ListItem
+                          onClick={() => {
+                            navigate(`/myprofile`);
+                          }}
+                        >
+                          회원정보
+                        </ListItem>
+                        <ListItem
+                          onClick={() => {
+                            onLogout();
+                          }}
+                        >
+                          로그아웃
+                        </ListItem>
+                      </DropDownList>
+                    </DropDownListContainer>
+                  )}
+                </DropDownContainer>
+
+                <NickBox
+                >{nickname}</NickBox>
               </>
             ) : (
               <>
-                <BtnBox>
+                <BtnBox2>
+                  <SearchIcon2
+                    onClick={() => {
+                      navigate(`/paper/search`);
+                    }}
+                  >
+                    <BiSearchAlt2 color="black" size="25px" />
+                  </SearchIcon2>
                   <BtnItem>
                     <Btn
                       onClick={() => {
@@ -168,7 +186,7 @@ const Header = () => {
                       로그인
                     </Btn>
                   </BtnItem>
-                </BtnBox>
+                </BtnBox2>
               </>
             )}
           </Login>
@@ -180,57 +198,63 @@ const Header = () => {
 
 const HeaderBox = styled.div`
   background-color: #fffdf7;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
   width: 100%;
-  height: 90px;
+  height: 143px;
+  border-bottom: 1px solid #A7ACA1;
 `;
 const Svg = styled.div`
   display: flex;
-  flex-direction: row;
+  width: 100%;
+  height: 143px;
+  position: relative;
   align-items: center;
-  margin: auto;
-`;
-const Logo = styled.div`
-  padding-left: 2%;
-  width: 27%;
-  height: 90px;
-  padding-left: 2%;
-  display: flex;
-  align-items: center;
-  background-color: #fffdf7;
-`;
-const Search = styled.div`
-  display: flex;
-  align-items: center !important;
   justify-content: center;
-  background-color: #fffdf7;
-  width: 46%;
-  height: 90px;
+  .paperTitle {
+    @media screen and (max-width: 800px) {
+    display: none;
+  }
+  }
 `;
+
 const Login = styled.div`
-  width: 27%;
-  height: 85px;
+  position: absolute;
+  right: 0px;
+  bottom: 20px;
+  width: 26%;
+  margin-right: 2.2%;
   display: flex;
   align-items: center;
   justify-content: end;
-  padding-right: 48px;
 `;
 const ProfileImgBox = styled.img`
   width: 40px;
   height: 40px;
-  margin: 0 0 0 0;
   border-radius: 50px;
   align-items: center;
 `;
-const BtnBox = styled.div`
-  width: 85%;
+const SearchIcon1 = styled.div`
   display: flex;
+  align-items: center;
+  margin-right: 10%;
+`
+const SearchIcon2 = styled.div`
+  display: flex;
+  align-items: center;
+`
+const BtnBox1 = styled.div`
+  width: 45%;
+  display: flex;
+  justify-content: end;
+  margin-right: 5%;
+`;
+const BtnBox2 = styled.div`
+  width: 78%;
+  display: flex;
+  justify-content: end;
   gap: 24px;
 `;
 const BtnItem = styled.div`
-  width: 80%;
+  width: 70%;
 `;
 
 const Btn = styled.button`
@@ -276,6 +300,10 @@ const ListItem = styled.li`
   cursor: pointer;
 `;
 const NickBox = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 82px;
   font-family: "Gmarket Sans";
   font-style: normal;
   font-weight: 400;
